refactor encryption service models
