+++ OutOfSchool/OutOfSchool.Common/Models/ExternalAuth/CertificateResponse.cs
@@ -1,0 +1,6 @@
+namespace OutOfSchool.Common.Models.ExternalAuth;
+
+public class CertificateResponse : IResponse
+{
+    public string CertBase64 { get; set; }
+}
