+++ OutOfSchool/OutOfSchool.Common/Models/ExternalAuth/EnvelopedUserInfoResponse.cs
@@ -1,0 +1,6 @@
+namespace OutOfSchool.Common.Models.ExternalAuth;
+
+public class EnvelopedUserInfoResponse : IResponse
+{
+    public string? EncryptedUserInfo { get; set; } = null;
+}
