--- OutOfSchool/OutOfSchool.Encryption/Models/CertificateResponse.cs
@@ -1,6 +1,0 @@
-namespace OutOfSchool.Encryption.Models;
-
-public class CertificateResponse
-{
-    public string CertBase64 { get; set; }
-}
