--- OutOfSchool/OutOfSchool.Encryption/Models/EnvelopedUserInfoResponse.cs
@@ -1,6 +1,0 @@
-namespace OutOfSchool.Encryption.Models;
-
-public class EnvelopedUserInfoResponse
-{
-    public string EncryptedUserInfo { get; set; }
-}
