+++ OutOfSchool/OutOfSchool.Encryption/Services/DevEUSignOAuth2Service.cs
@@ -1,14 +1,12 @@
 using System.Diagnostics.CodeAnalysis;
-using System.Text.Json;
-using OutOfSchool.Encryption.Models;
+using OutOfSchool.Common;
+using OutOfSchool.Common.Models.ExternalAuth;
 
 namespace OutOfSchool.Encryption.Services;
 
 /// <inheritdoc/>
 public class DevEUSignOAuth2Service : IEUSignOAuth2Service
 {
-    private readonly JsonSerializerOptions jsonSerializerOptions = new(JsonSerializerDefaults.Web);
-
     /// <inheritdoc/>
     public CertificateResponse GetEnvelopeCertificateBase64() => new()
     {
@@ -24,6 +22,6 @@ public class DevEUSignOAuth2Service : IEUSignOAuth2Service
         }
 
         // Mock local auth server sends data as unencrypted string.
-        return JsonSerializer.Deserialize<UserInfoResponse>(encryptedUserInfo.EncryptedUserInfo, jsonSerializerOptions);
+        return JsonSerializerHelper.Deserialize<UserInfoResponse>(encryptedUserInfo.EncryptedUserInfo);
     }
 }
